Update mutations to use built in ID field
Graphql/Relay generates a ID field for every non-scalar type. We
were using a uuid field to identify objects for our mutations, but
the queries were using the ID field. This commit replaces all usages
of UUIDs with the standard graphql ID field. This is possible using
the `from_global_id` method in `graphql_relay`, which is sadly
pretty much undocumented. It converts from the global id to a
django model id tuple which contains the model name and a public key
(e.g `('User', 1)`).